feat(graph)!: an identity is a bigint, because a binding does not save a 53-bit number
`VertexId` was `number & brand` and the pair was packed by `type * 2**32 + dense`. Both are now
64-bit: `bigint` at the boundary, `BigUint64Array` in the slice, shift-and-mask instead of
`Math.floor`.

The evidence is not ours and is worth carrying in the comments. Across Arrow, Parquet, Iceberg,
Delta, MVT, PMTiles, Zarr, GraphAr, H3 and S2, the failure that recurs is a 64-bit value crossing
into JavaScript. `mapbox/node-s2` is a *binding* to the same C++ the reference implementation runs
and it still returned wrong cell ids — issue #92, open since 2017, `1152921504606847000` where Java
and Go give `1152921504606846977`. A binding cannot protect a language boundary that cannot hold the
value. H3 settled it by decree before anyone could get it wrong: `h3-js` types `H3Index` as a
string. rmlext ADR-0045 §4 takes the same decision with the type JavaScript grew for it.

Two things fall out that the brand alone could not buy. A buffer index is a `number` and an identity
is a `bigint`, so confusing them stops being a convention and becomes a primitive type error that no
cast quietly launders — `7 as VertexId` compiled before and does not now, and the remaining hole is
`as unknown as`, which is loud. And `>>` in JavaScript converts to 32 bits and takes the count modulo
32, so `type << 32 | dense` on a `number` is not a lost high word, it is `type | dense`: the packing
cannot be written correctly on a `number` at all.

No lint rule was added. The ADR asked for one banning `number` for an id; with `bigint` the compiler
already rejects it at every call site, and a rule that restates a type error is a second thing to
keep in sync. What a rule could not have caught either is the cast, so that is a test instead:
`resident.test.ts` carries two `@ts-expect-error`s, one per layer — a `number` for the primitive, a
bare `bigint` for the brand — and removing either makes `pnpm typecheck` red on an unused
expectation. Also new there: the far corner of the 64-bit range round-trips through the buffer,
which the old packing got wrong by 616 rather than by an error, and `Map`/`Set` are shown to key
bigints by value rather than assumed to.

`fillColumn` writes `Number(…)` and a `BigUint64Array` element takes a `bigint` only, so
`duck-source` fills a scratch and widens on the way across. That refusal is the guarantee, so the
extra buffer is the price of the boundary being enforced by the runtime instead of by us.

Nothing changed on the way to the GPU: positions and buffer indices are still `number` and
`Float32Array`. Verified live over the workspace showcase with its limit forced to 120 so the
1,543-node corpus is sliced: `slice.vertices` is a `BigUint64Array` of 120, a marquee of six
resolved to identities `[7, 23, 24, 41, 42, 47]` at indices `[7, 23, 24, 41, 42, 46]`, one survivor
after a pan at index 12, and the original six indices again on the way back — the selection never
touched. The analytics subpath measures 66.28 kB against its 68 kB limit, unchanged; the graph
bundle grew 17 bytes brotlied, because BigInt is a primitive and not a polyfill.

+++ packages/graph/src/bounded.ts
@@ -98,11 +98,13 @@ export interface Slice {
    * to be held as an identity and re-resolved through `residentOf` each time. Leaving this out was
    * how the first draft would have shipped a selection that silently pointed at the wrong nodes.
    *
-   * `Float64Array` rather than `Uint32Array` because the pair does not fit in 32 bits, and a dense id
-   * on its own is not an identity: it numbers within one vertex type, so a union of two types repeats
-   * every value.
+   * `BigUint64Array` because the pair is 64 bits exactly — a `Uint32Array` cannot hold it at all,
+   * and a `Float64Array` holds it only while the type index stays under 2²¹, which is a ceiling
+   * nobody would find until they crossed it. Still a typed array, so it is still one allocation and
+   * still transferable; only what it carries changed. A dense id on its own is not an identity: it
+   * numbers within one vertex type, so a union of two types repeats every value.
    */
-  vertices: Float64Array;
+  vertices: BigUint64Array;
   /** `[x0, y0, x1, y1, …]`, one pair per returned point. */
   positions: Float32Array;
   /** `[src, dst, …]` as indices into `positions`. */
